Quick comments
Also if you read this when loading the unity project of the render pipelines are having errors and I need to work on my other projects so the nav mesh for the main area might be a mess. Also everything is in the Playground scene so I deleted the sample scene.